+++ tf2rl/experiments/irl_trainer.py
@@ -90,7 +90,7 @@ class IRLTrainer(Trainer):
                     rew = self._irl.inference(samples["obs"], samples["act"])
                     td_error = self._policy.train(
                         samples["obs"], samples["act"], samples["next_obs"],
-                        rew, np.array(samples["done"], dtype=np.float64),
+                        rew, np.array(samples["done"], dtype=np.float32),
                         None if not self._use_prioritized_rb else samples["weights"])
                     if self._use_prioritized_rb:
                         replay_buffer.update_priorities(samples["indexes"], np.abs(td_error) + 1e-6)
